requirements.txt
anyio==3.6.2
black==23.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
fastapi==0.89.1
fastapi-security==0.5.0
h11==0.14.0
idna==3.4
jose==1.0.0
jwt==1.3.1
mypy-extensions==0.4.3
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==2.6.2
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
